requirements.txt
fastapi
uvicorn
asyncio
psycopg2-binary
pydantic
uvicorn